Extend ActiveSupport::TimeWithZone the same way we extend Time
