+++ lib/business_time.rb
@@ -7,6 +7,8 @@ require 'business_time/config'
 require 'business_time/business_hours'
 require 'business_time/business_days'
 require 'business_time/core_ext/date'
-require 'business_time/core_ext/time'
 require 'business_time/core_ext/fixnum'
 
+require 'business_time/time_extensions'
+require 'business_time/core_ext/time'
+require 'business_time/core_ext/active_support/time_with_zone'
