+++ lib/business_time/core_ext/active_support/time_with_zone.rb
@@ -1,0 +1,3 @@
+class ActiveSupport::TimeWithZone
+  include BusinessTime::TimeExtensions
+end
